requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose==3.3.0
bcrypt==4.1.3
pydantic==2.6.3
python-multipart==0.0.9
